将要 更新UI，分开上传、下载的管理

+++ SmbFileExplorer/KxSMBProvider.h
@@ -73,7 +73,7 @@ typedef enum {
 @class KxSMBItem;
 
 typedef void (^KxSMBBlock)(id result);
-typedef void (^KxSMBBlockProgress)(KxSMBItem *item, long transferred);
+typedef void (^KxSMBBlockProgress)(KxSMBItem *item, unsigned long long transferred);
 
 @interface KxSMBItemStat : NSObject
 @property(readonly, nonatomic, strong) NSDate *lastModified;
